requirements.txt
requests
pandas
supernotelib
ocrmypdf
transformers 
torch 
torchvision 
pillow
python-docx
easyocr